Raise an IOError if connection to socket fails.
Following the statement "We try to catch them and re-throw Python standard exceptions." from the beginning of the file, it makes sense to raise some standard exception when no connection to the socket can be established. The standard python exception that fits best (imo) is IOError as this is also the parent of socket.error.

Note: It might make more sense to raise a socket.error exception instead but that would require importing socket.

Note #2: I didn't write a test for this case since I see no tests related to raising exceptions in the test suit (those could of course be added).